refactor: move profile folder and components to screens folder
